requirements_dev.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements_dev.in
#
appdirs==1.4.4
    # via black
asgiref==3.5.0
    # via
    #   -r requirements.txt
    #   django
asttokens==2.0.5
    # via stack-data
attrs==21.4.0
    # via black
babel==2.10.1
    # via -r requirements_dev.in
backcall==0.2.0
    # via ipython
beautifulsoup4==4.10.0
    # via
    #   -r requirements.txt
    #   django-bootstrap4
black==19.10b0
    # via
    #   -r requirements_dev.in
    #   ipython
certifi==2021.10.8
    # via
    #   -r requirements.txt
    #   requests
cffi==1.15.0
    # via
    #   -r requirements.txt
    #   cryptography
charset-normalizer==2.0.12
    # via
    #   -r requirements.txt
    #   requests
click==8.0.4
    # via
    #   black
    #   pip-tools
coverage==6.3.2
    # via -r requirements_dev.in
cryptography==36.0.1
    # via
    #   -r requirements.txt
    #   jwcrypto
    #   pyjwt
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
defusedxml==0.7.1
    # via
    #   -r requirements.txt
    #   python3-openid
deprecated==1.2.13
    # via
    #   -r requirements.txt
    #   jwcrypto
django==3.2.13
    # via
    #   -r requirements.txt
    #   django-admin-sortable2
    #   django-allauth
    #   django-axes
    #   django-bootstrap-datepicker-plus
    #   django-bootstrap4
    #   django-ckeditor
    #   django-common-helpers
    #   django-cors-headers
    #   django-cron
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-formtools
    #   django-js-asset
    #   django-oauth-toolkit
    #   django-otp
    #   django-phonenumber-field
    #   django-picklefield
    #   django-polymorphic
    #   django-ranged-response
    #   django-simple-captcha
    #   django-tables2
    #   django-tables2-column-shifter
    #   django-taggit
    #   django-two-factor-auth
    #   djangorestframework
django-admin-sortable2==1.0.4
    # via -r requirements.txt
django-allauth==0.45.0
    # via -r requirements.txt
django-axes==5.31.0
    # via -r requirements.txt
django-bootstrap-datepicker-plus==4.0.0
    # via -r requirements.txt
django-bootstrap4==21.2
    # via -r requirements.txt
django-ckeditor==6.4.2
    # via -r requirements.txt
django-common-helpers==0.9.2
    # via
    #   -r requirements.txt
    #   django-cron
django-constance[database]==2.8.0
    # via -r requirements.txt
django-cors-headers==3.11.0
    # via -r requirements.txt
django-crispy-forms==1.14.0
    # via -r requirements.txt
django-cron==0.5.1
    # via -r requirements.txt
django-debug-toolbar==3.2.4
    # via -r requirements_dev.in
django-extensions==3.1.5
    # via -r requirements_dev.in
django-filter==21.1
    # via -r requirements.txt
django-formtools==2.3
    # via
    #   -r requirements.txt
    #   django-two-factor-auth
django-grappelli==3.0.3
    # via -r requirements.txt
django-ipware==4.0.2
    # via
    #   -r requirements.txt
    #   django-axes
django-js-asset==2.0.0
    # via
    #   -r requirements.txt
    #   django-ckeditor
django-oauth-toolkit==1.7.0
    # via -r requirements.txt
django-otp==1.1.3
    # via
    #   -r requirements.txt
    #   django-two-factor-auth
django-phonenumber-field==6.1.0
    # via
    #   -r requirements.txt
    #   django-two-factor-auth
django-picklefield==3.0.1
    # via
    #   -r requirements.txt
    #   django-constance
django-polymorphic==3.1.0
    # via -r requirements.txt
django-ranged-response==0.2.0
    # via
    #   -r requirements.txt
    #   django-simple-captcha
django-simple-captcha==0.5.17
    # via -r requirements.txt
django-simple-history==3.0.0
    # via -r requirements.txt
django-tables2==2.4.1
    # via
    #   -r requirements.txt
    #   django-tables2-column-shifter
django-tables2-column-shifter==2.0.3
    # via -r requirements.txt
django-taggit==2.1.0
    # via -r requirements.txt
django-two-factor-auth[phonenumbers]==1.13.2
    # via -r requirements.txt
djangorestframework==3.13.1
    # via
    #   -r requirements.txt
    #   djangorestframework-gis
djangorestframework-gis==0.18
    # via -r requirements.txt
docker==5.0.3
    # via -r requirements.txt
executing==0.8.2
    # via stack-data
factory-boy==3.2.1
    # via -r requirements_dev.in
faker==13.2.0
    # via
    #   -r requirements_dev.in
    #   factory-boy
filetype==1.0.10
    # via -r requirements.txt
gdal==3.3.2
    # via -r requirements.txt
gunicorn==20.1.0
    # via -r requirements.txt
html5lib==1.1
    # via -r requirements.txt
idna==3.3
    # via
    #   -r requirements.txt
    #   requests
ipdb==0.13.9
    # via -r requirements_dev.in
ipython==8.0.1
    # via ipdb
jdcal==1.4.1
    # via -r requirements.txt
jedi==0.18.1
    # via ipython
jinja2==3.1.2
    # via -r requirements.txt
jwcrypto==1.0
    # via
    #   -r requirements.txt
    #   django-oauth-toolkit
markupsafe==2.1.1
    # via
    #   -r requirements.txt
    #   jinja2
matplotlib-inline==0.1.3
    # via ipython
oauthlib==3.2.0
    # via
    #   -r requirements.txt
    #   django-oauth-toolkit
    #   requests-oauthlib
parso==0.8.3
    # via jedi
pathspec==0.9.0
    # via black
pdf2image==1.16.0
    # via -r requirements.txt
pep517==0.12.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
phonenumbers==8.12.43
    # via
    #   -r requirements.txt
    #   django-two-factor-auth
pickleshare==0.7.5
    # via ipython
pillow==9.0.1
    # via
    #   -r requirements.txt
    #   django-simple-captcha
    #   pdf2image
pip-tools==6.5.1
    # via -r requirements_dev.in
prompt-toolkit==3.0.28
    # via ipython
psycopg2-binary==2.9.3
    # via -r requirements.txt
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via
    #   -r requirements.txt
    #   cffi
pygments==2.11.2
    # via ipython
pyjwt[crypto]==2.3.0
    # via
    #   -r requirements.txt
    #   django-allauth
python-dateutil==2.8.2
    # via faker
python-dotenv==0.19.2
    # via -r requirements.txt
python3-openid==3.2.0
    # via
    #   -r requirements.txt
    #   django-allauth
pytz==2021.3
    # via
    #   -r requirements.txt
    #   babel
    #   django
    #   djangorestframework
pyyaml==6.0
    # via -r requirements.txt
qrcode==7.3.1
    # via
    #   -r requirements.txt
    #   django-two-factor-auth
regex==2022.1.18
    # via black
requests==2.27.1
    # via
    #   -r requirements.txt
    #   django-allauth
    #   django-oauth-toolkit
    #   docker
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via
    #   -r requirements.txt
    #   django-allauth
six==1.16.0
    # via
    #   -r requirements.txt
    #   html5lib
    #   python-dateutil
soupsieve==2.3.1
    # via
    #   -r requirements.txt
    #   beautifulsoup4
sqlparse==0.4.2
    # via
    #   -r requirements.txt
    #   django
    #   django-debug-toolbar
stack-data==0.2.0
    # via ipython
tablib==3.2.0
    # via -r requirements.txt
toml==0.10.2
    # via
    #   black
    #   ipdb
tomli==2.0.1
    # via pep517
traitlets==5.1.1
    # via
    #   ipython
    #   matplotlib-inline
typed-ast==1.5.2
    # via black
uritemplate==4.1.1
    # via -r requirements.txt
urllib3==1.26.8
    # via
    #   -r requirements.txt
    #   requests
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   -r requirements.txt
    #   html5lib
websocket-client==1.3.2
    # via
    #   -r requirements.txt
    #   docker
wheel==0.37.1
    # via pip-tools
whitenoise==6.0.0
    # via -r requirements.txt
wrapt==1.13.3
    # via
    #   -r requirements.txt
    #   deprecated

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
